Fix failed tests
ageの型変更に伴ない、失敗していたCustomer2クラスのテストを修正

+++ OCMockTests/Customer2Test.m
@@ -41,52 +41,52 @@
 
 - (void)testAge_0{
     [_sut setBirthWithString:@"2013/3/2"];
-    XCTAssertEqual(0U, _sut.age);   //ageはunsignedなので、expectedにもUを付加
+    XCTAssertEqual(0U, (NSUInteger)_sut.age);   //ageはunsignedなので、expectedにもUを付加
 
     [_sut setBirthWithString:@"2014/3/1"];
-    XCTAssertEqual(0U, _sut.age);
+    XCTAssertEqual(0U, (NSUInteger)_sut.age);
 
     [_sut setBirthWithString:@"2014/3/2"];
-    XCTAssertEqual(0U, _sut.age);
+    XCTAssertEqual(0U, (NSUInteger)_sut.age);
 }
 
 - (void)testAge_1{
     [_sut setBirthWithString:@"2013/3/1"];
-    XCTAssertEqual(1U, _sut.age);
+    XCTAssertEqual(1U, (NSUInteger)_sut.age);
 
     [_sut setBirthWithString:@"2012/3/2"];
-    XCTAssertEqual(1U, _sut.age);
+    XCTAssertEqual(1U, (NSUInteger)_sut.age);
 
     [_sut setBirthWithString:@"2012/3/1"];
-    XCTAssertEqual(2U, _sut.age);
+    XCTAssertEqual(2U, (NSUInteger)_sut.age);
 }
 
 - (void)testAge_17{
     [_sut setBirthWithString:@"1997/3/2"];
-    XCTAssertEqual(16U, _sut.age);
+    XCTAssertEqual(16U, (NSUInteger)_sut.age);
 
     [_sut setBirthWithString:@"1997/3/1"];
-    XCTAssertEqual(17U, _sut.age);
+    XCTAssertEqual(17U, (NSUInteger)_sut.age);
 
     [_sut setBirthWithString:@"1996/3/2"];
-    XCTAssertEqual(17U, _sut.age);
+    XCTAssertEqual(17U, (NSUInteger)_sut.age);
 
     [_sut setBirthWithString:@"1996/3/1"];
-    XCTAssertEqual(18U, _sut.age);
+    XCTAssertEqual(18U, (NSUInteger)_sut.age);
 }
 
 - (void)testAge_100{
     [_sut setBirthWithString:@"1914/3/2"];
-    XCTAssertEqual(99U, _sut.age);
+    XCTAssertEqual(99U, (NSUInteger)_sut.age);
 
     [_sut setBirthWithString:@"1914/3/1"];
-    XCTAssertEqual(100U, _sut.age);
+    XCTAssertEqual(100U, (NSUInteger)_sut.age);
 
     [_sut setBirthWithString:@"1913/3/2"];
-    XCTAssertEqual(100U, _sut.age);
+    XCTAssertEqual(100U, (NSUInteger)_sut.age);
 
     [_sut setBirthWithString:@"1913/3/1"];
-    XCTAssertEqual(101U, _sut.age);
+    XCTAssertEqual(101U, (NSUInteger)_sut.age);
 }
 
 @end
